+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 - Added play all/shuffle buttons to main screen and albums
 - Enhanced the search route by adding filtering by content category
 - Now all queues are saved/restored. Prior in a lot of cases the queue was intentionally not restored
+- Album origins are now restored, that means if you queue some album, after app reload when song from
+  this album will be played, the playing indicator will indicate the album as playing. Prior, the origin
+  was saved only for the current app session, when it was added, and was not restored.
 - Optimized scrollbar and listviews
 - Became media browser service, support for Android Auto
-- Refactored a lot of code and fixed a lot of bugs
+- Fixed that album arts were not recahced (below Android 11)
+- Refactored a lot of code and fixed a lot of other bugs
 - Changed player backend (non-UX)
 
 ## [1.0.1] - february 08, 2021